chore: AnswerStats 내 count 타입을 Long으로 변경 및 로그 추가

+++ src/main/java/OneQ/OnSurvey/domain/participation/model/dto/AnswerStats.java
@@ -7,11 +7,11 @@ import lombok.Getter;
 public class AnswerStats {
     Long questionId;
     String content;
-    Integer count;
+    Long count;
 
     public AnswerStats(Long questionId, String content) {
         this.questionId = questionId;
         this.content = content;
-        this.count = 1;
+        this.count = 1L;
     }
 }
